+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix pool `execute()` arguments check.
+
 ### Changed
 
 - Make continuous tasks stealing algorithm less aggressive.